wip: transitioned to new approach
new approach creates an object for each layer of load stack with a unique name which can then be merged onto load data. page will end up with all the data from each step (all layouts, and page) and then can merge them on page as opposed to in load functions. this allows us to run load functions in parallel and get better performance and cleaner syntax

also improved the api of the tool with more flexibility now